+++ src/main_algorithm.jl
@@ -282,7 +282,11 @@ end
    A wrapper function that collects automated solver adjustments within the main while loop.
 """
 function algorithm_automation(m::Optimizer)
-    Alp.get_option(m, :disc_var_pick) == 3 && Alp.update_disc_cont_var(m)
+    
+    # Only if a feasible solution is detected
+    if (Alp.get_option(m, :disc_var_pick) == 3) && m.detected_incumbent
+        Alp.update_disc_cont_var(m)
+    end
 
     if Alp.get_option(m, :partition_scaling_factor_branch)
         Alp.set_option(